yorch-backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI
google-generativeai==0.8.3

# Cloudinary
cloudinary==1.42.0

# Utils
python-dotenv==1.0.1
httpx==0.28.1
